Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -3,7 +3,7 @@
 int main() {
     long long int n;
     scanf("%lld",n)  
-    int msb = (n >> 31) & 1;  
+    long long int msb = (n >> 31) & 1;  
     if (msb == 1) {
         printf("Set");
     } else {
